This test used to take ~5s. Now it's ~0.3s.

Remaining slowest test is `azkaban.execapp.JobRunnerTest`: 2.513s.

All enabled tests now run in ~5s \o/