@HappyRay 2017-05-26
Excellent. Thanks!

LGTM. Could you update the branch or do you mind that I update the branch so that we can merge?

It appears that you shortened the various Jetty timeout values. It would be great if you can describe what you did to speed up the tests.