@juhoautio 2017-05-26
Updated branch.

How this was speeded up: This test was being run using the default timeout values. In the test the normal request time is so short that we can set lower timeouts in test props and still test the timeout scenarios reliably.